+++ lib/tappay/errors.rb
@@ -4,7 +4,4 @@ module Tappay
   class ConfigurationError < Error; end
   class ConnectionError < Error; end
   class ValidationError < Error; end
-  class PaymentError < Error; end
-  class RefundError < Error; end
-  class QueryError < Error; end
 end